Added hotplug rule for USB hub
When turning off power to D-Link DUB-H4 sometimes it likes to disconnect from the kernel and immediately reconnect causing the fans to come back on. Addition of a hotplug usb rule for the device ensures uhubctl stops them again if this happens. Without this the fan state was not truly known (i.e. sometimes it would start without uhubctl explicitly telling it to start and run until the next cron scheduled run of the script).

Even just querying the hub state with uhubctl or 'lsusb -v' would sometimes cause the hub to disconnect/reconnect (which is why we now use $FANON file instead of grepping the active status). After much testing and RTFM'ing, it seems this is just a quirk with either this particular hub or the Linux kernel. Addition of the hotplug rule seems to be the most consistent and straightforward workaround for this issue.