Moved modem restart to own function
Also added logic to properly perform a full modem reset if $LIFACE is cycled. The RM502Q-AE appears to need a full restart when it loses cell connectivity else NSA CA abilities don't kick in on the re-established connection. The default ModemManager only disables the modem before re-attaching so we detect the modem disable and force a full restart to be sure we get proper speeds.